Rename variables.py into constants.py
Handle configuration loading actions.
