+++ workers/arduino/arduino_control_worker.py
@@ -14,7 +14,7 @@ import sys
 
 
 
-import variables
+import constants
 import importlib
 from logger.Logger import Logger, LOG_LEVEL
