+++ workers/sequence_worker.py
@@ -8,7 +8,7 @@ import sys
 
 
 
-import variables
+import constants
 import importlib
 
 from logger.Logger import Logger, LOG_LEVEL
