+++ workers/worker.py
@@ -7,7 +7,7 @@ import sys
 
 
 
-import variables
+import constants
 from logger.Logger import Logger, LOG_LEVEL
